Module r_freertos::ffi::queue

Queue ffi for C

Functions

cGetQueueRxLock

Get RxLock of target queue handle.

cGetQueueTxLock

Get TxLock of target queue handle.

uxQueueMessagesWaiting

Get uxMessagesWaiting of target queue handle.

uxQueueSpacesAvailable

Get available space of target queue handle.

vQueueDeleteToC

The C version of vQueueDelete.

vSetQueueRxLock

Set RxLock of target queue handle.

vSetQueueTxLock

Set TxLock of target queue handle.

xQueueCreateToC

The C version of xQueueCreate.

xQueueGenericSendToC

The C version of xQueueGenericSend.
The implement is slightly different from rust version while the function is same.

xQueueIsQueueEmptyFromISR

The ISR version of prvIsQueueEmpty for C.
Not impleted in rust.

xQueueIsQueueFullFromISR

The ISR version of prvIsQueueFull for C.
Not impleted in rust.

xQueuePeekFromISRToC

The C version of xQueuePeekFromISR.

xQueuePeekToC

The C version of xQueueReceive.
The implement is slightly different from rust version while the function is same.

xQueueReceiveFromISRToC

The C version of xQueueReceiveFromISR.

xQueueReceiveToC

The C version of xQueueReceive.
The implement is slightly different from rust version while the function is same.

xQueueResetToC

The C version of xQueueReset.

xQueueSendFromISRToC

The C version of xQueueSendFromISR.

xQueueSendToC

The C version of xQueueSend.

Type Definitions

QueueHandle_c