Module r_freertos::kernel::linked_list

Bidirectional linked list Definition and API

Structs

XList

XListItem

Functions

list_current_list_length

Get num of item in list.

list_get_c_owner_of_head_entry

Get owner of head entry in list.
Owner is saved as C ptr.
Does not alter current index.

list_get_end_marker

Get end marker of list.
The end marker is not a valid item.

list_get_head_entry

Get head entry of list.
Head entry is the first valid item in the list, unless the list is empty.
In that case, the end marker will be returned.

list_get_num_items

Get num of item in list.

list_get_owner_of_head_entry

Get owner of head entry in list.
Does not alter current index.

list_get_owner_of_next_entry

Get owner of next entry in list.
Move current index to next item.

list_get_pxindex

Get current index of list.

list_get_value_of_head_entry

Return the value of list’s head entry.

list_is_contained_within

Return if the list item is contained within target list.

list_is_empty

Return if the list is empty.

list_item_get_c_owner

Get owner of item.
Owner is saved as C ptr.

list_item_get_container

Get container of item.

list_item_get_next

Get next item of target item.

list_item_get_owner

Get owner of item.
Owner is a tskTaskControlBlock object.

list_item_get_pre

Get previous item of target item.

list_item_get_value

Get value of item.
Insert() place item in order of values.

list_item_set_container

Set container of item.

list_item_set_next

Set next item of target item.

list_item_set_owner

Set container of item.

list_item_set_pre

Set previous item of target item.

list_item_set_value

Set value of item.
Insert() place item in order of values.

list_set_pxindex

Set current index of list.

ux_list_remove

Remove target item from its container.
Return number of items after remove.

v_list_insert

Insert target item into list in ascending order of value.
If target item has value==PORT_MAX_DELAY, insert to list end.
If list is not already in order, insert position is not guaranteed.

v_list_insert_end

Rnsert target item into end of list.

Type Definitions

ListItemLink

ListItemOwnerWeakLink

ListItemT

ListItemWeakLink

ListRealLink

ListT

ListWeakLink