Function r_freertos::kernel::linked_list::v_list_insert

pub fn v_list_insert(
    px_list: &Arc<RwLock<XList>>, 
    px_new_list_item: &Arc<RwLock<XListItem>>
)

Insert target item into list in ascending order of value.
If target item has value==PORT_MAX_DELAY, insert to list end.
If list is not already in order, insert position is not guaranteed.