Module r_freertos::kernel::linked_list

linked list implementation and api

Structs

XList

XListItem

Functions

list_current_list_length

get num of item in list(Arc)

list_get_end_marker

get end marker of list
end marker is not a valid item

list_get_head_entry

get head entry of list
head entry is the first valid item in the list, unless the list is empty

list_get_num_items

get num of item in list(Weak)

list_get_owner_of_head_entry

get owner of head entry in list
do not alter current index

list_get_owner_of_next_entry

get owner of next entry in list
move current index to next item

list_get_pxindex

get current index of list

list_is_empty

return if the list is empty

list_item_get_container

get container of item

list_item_get_next

get next item

list_item_get_owner

get owner of item
owner is a tskTaskControlBlock object

list_item_get_pre

get previous item

list_item_get_value

get value of item
insert() place item in order of values

list_item_set_container

set container of item

list_item_set_next

set next item

list_item_set_owner

set container of item

list_item_set_pre

set previous item

list_item_set_value

set value of item
insert() place item in order of values

list_set_pxindex

set current index of list

ux_list_remove

remove target item from its container

v_list_insert

insert target item into list in ascending order of value
if target item has value==PORT_MAX_DELAY, insert to list end
if list is not already in order, insert position is not guaranteed

v_list_insert_end

insert target item into end of list

Type Definitions

ListItemLink

ListItemOwnerWeakLink

ListItemT

ListItemWeakLink

ListRealLink

ListT

ListWeakLink